# back-end/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The official documentation of StudyTogether",
    "version": "1.0.5",
    "title": "StudyTogether's Offical API Documentation"
  },
...
  "paths": {
    "/users/login": {
      "post": {
        "tags": ["User"],
        "summary": "Logs in exising user",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "description": "The user name for login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/credentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "headers": {
              "token": {
                "type": "string",
                "description": "JWT token for the user"
              }
            }
          },
          "400": {
            "description": "Invalid username or password"
          },
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "User was not found in records"
          },
          "401": {
            "description": "Password was invalid"
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": ["User"],
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "string",
                "description": "JWT token for the user"
              }
            }
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/profile/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get profile information. Amount of data returned back depends on whether the user is request their profile or another user's profile.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/PublicProfile"
            }
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/profile/follow/{id}": {
      "patch": {
        "tags": ["User"],
        "summary": "Adds the requesting user as a follower if user {id}. It also adds updates the following list of the requesting user",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/profile/unfollow/{id}": {
      "patch": {
        "tags": ["User"],
        "summary": "undoes the actions of the follow endpoint",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/profile/{id}/followers": {
      "get": {
        "tags": ["User"],
        "summary": "Returns followers of user {id}",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/profile/{id}/following": {
      "get": {
        "tags": ["User"],
        "summary": "Returns users who {id} followes.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/notifications": {
      "get": {
        "tags": ["User"],
        "summary": "Returns notifications relevant to the user",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/profile/recommend/{limit}": {
      "get": {
        "tags": ["User"],
        "summary": "Returns a list of users the requester has not followed. The amount of users in this list is bound by {limit}",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "User identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid parameter(s) supplied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/bookmark-group": {
      "post": {
        "tags": ["User"],
        "summary": "Bookmark a study group by id",
        "description": "",
        "parameters": [
          {
            "name": "studygroupId",
            "in": "body",
            "description": "ID of study group to bookmark",
            "required": true,
            "type": "string",
            "example": "{\"studygroupId\": \"62034213b10d3c4264eb84c7\"}"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "Study group was bookmarked successfully"
              }
            }
          },
          "400": {
            "description": "Study group with the given id was already bookmarked"
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "404": {
            "description": "Study group with the given ID was not found in the database"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/unbookmark-group": {
      "patch": {
        "tags": ["User"],
        "summary": "Unbookmark a study group by id",
        "description": "",
        "parameters": [
          {
            "name": "studygroupId",
            "in": "body",
            "description": "ID of study group to bookmark",
            "required": true,
            "type": "string",
            "example": "{\"studygroupId\": \"62034213b10d3c4264eb84c7\"}"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "format is 'JWT ' followed by the token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "token": {
                "type": "json",
                "description": "Study group was unbookmarked successfully"
              }
            }
          },
          "400": {
            "description": "Study group with the given id has not been bookmarked"
          },
          "401": {
            "description": "Unauthenticated user"
          },
          "404": {
            "description": "Study group with the given ID was not found in the database"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "user": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Tobey",
          "required": true
        },
        "lastName": {
          "type": "string",
          "example": "Maguire",
          "required": true
        },
        "userName": {
          "type": "string",
          "example": "SpiderMan"
        },
        "email": {
          "type": "string",
          "example": "Tobey.Maguire@mail.utoronto.ca",
          "required": true
        },
        "password": {
          "type": "string",
          "example": "ddf2834DFG$^^%xjfhggf#$%^",
          "required": true
        },
        "role": {
          "type": "string",
          "example": "Student",
          "required": true
        },
        "verified": {
          "type": "boolean",
          "example": false,
          "required": true,
          "defualt": false
        },
        "created": {
          "type": "date",
          "example": "2022-11-07T17:04:15.000+00:00",
          "required": true
        }
      }
    },
    "PublicProfile": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Tobey",
          "required": true
        },
        "lastName": {
          "type": "string",
          "example": "Maguire",
          "required": true
        },
        "userName": {
          "type": "string",
          "example": "SpiderMan"
        },

        "role": {
          "type": "string",
          "example": "Student",
          "required": true
        }
      }
    },
    "credentials": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "Tobey.Maguire@mail.utoronto.ca",
          "required": true
        },
        "password": {
          "type": "string",
          "example": "ddf2834DFG$^^%xjfhggf#$%^",
          "required": true
        }
      }
    },
...
  }
}